+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- Fix for the undo manager so it will correctly group reference invalidations inside the action that caused it.
 - `ExtendedModel` will now throw when the `baseModel` argument is not a model class.
 
 ## 0.48.3